floor a float

+++ 0x00-python_variable_annotations/2-floor.py
@@ -1,12 +1,11 @@
 #!/usr/bin/env python3
-
 """
 Returns floor of a float
 """
 import math
 
 
-def floor(n: float) -> float:
+def floor(n: float) -> int:
     """Returns the floor of the argument
 
     Args:
